Fixes compilation error caused by the esp32 board package update in 2.0.3. It seems that one has to use the exact type now to resolve the typechecking.

+++ Ezo_i2c.cpp
@@ -38,7 +38,7 @@ void Ezo_board::send_read_cmd(){
 }
 
 void Ezo_board::send_cmd_with_num(const char* cmd, float num, uint8_t decimal_amount){
-	String temp = String(cmd )+ String(num, decimal_amount);
+	String temp = String(cmd )+ String(num, (unsigned int)decimal_amount);
 	const char* pointer = temp.c_str();
 	send_cmd(pointer);
 }
